KinesisFireHoseWithS3Bucket/requirements.txt
-e .
aws-cdk.aws_s3
aws_cdk.aws_kinesisfirehose